code/requirements.txt
asgiref==3.8.1
attrs==25.3.0
black==25.1.0
click==8.1.8
coverage==7.8.0
dependency-injector==4.46.0
Django==4.2.20
django-seed==0.3.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
drf-spectacular==0.28.0
drf-spectacular-sidecar==2025.4.1
Faker==37.1.0
flake8==7.2.0
inflection==0.5.1
isort==6.0.1
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
mccabe==0.7.0
mypy==1.15.0
mypy_extensions==1.1.0
packaging==25.0
pathspec==0.12.1
platformdirs==4.3.7
psycopg2-binary==2.9.10
pycodestyle==2.13.0
pyflakes==3.3.2
PyJWT==2.9.0
python-dotenv==1.1.0
PyYAML==6.0.2
referencing==0.36.2
rpds-py==0.24.0
sqlparse==0.5.3
toposort==1.10
typing_extensions==4.13.2
tzdata==2025.2
uritemplate==4.1.1
